refactor: split helper functions into seperate 'helpers.py' script
